requirements.txt
# python=3.8.8

fastapi
nltk==3.6.3
numpy<1.24
pandas
python-dotenv
scikit-learn==0.22.1
uvicorn[standard]
